@@ -4,6 +4,9 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
+    "clean": "rm -rf node_modules",
+    "reinstall": "npm run clean && npm install",
+    "rebuild": "npm run clean && npm install && npm run build",
     "start": "ng build && node ./bin/www",
     "build": "ng build --aot --prod",
     "test": "ng test",
@@ -12,16 +15,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^6.0.0",
+    "@angular/animations": "^6.1.10",
+    "@angular/common": "^6.1.10",
     "@angular/compiler": "^6.1.10",
-    "@angular/compiler-cli": "^6.1.10",
-    "@angular/core": "^6.0.0",
-    "@angular/forms": "^6.0.0",
-    "@angular/http": "^6.0.0",
+    "@angular/core": "^6.1.10",
+    "@angular/forms": "^6.1.10",
+    "@angular/http": "^6.1.10",
     "@angular/platform-browser": "^6.1.10",
     "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/router": "^6.0.0",
+    "@angular/router": "^6.1.10",
     "angular-bootstrap-md": "^6.2.5",
     "angular5-social-login": "^1.0.9",
     "bootstrap": "^4.0.0-beta.3",
@@ -32,17 +34,18 @@
     "hammerjs": "^2.0.8",
     "jquery": "^3.3.1",
     "mongodb": "^3.0.3",
-    "mysql": "^2.15.0",
+    "mysql": "^2.16.0",
     "nodemon": "^1.18.4",
     "popper.js": "^1.12.9",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.2.2",
     "rxjs-compat": "^6.3.3",
+    "rxjs-tslint": "^0.1.5",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.8.5",
-    "@angular/cli": "^6.2.5",
-    "@angular/compiler-cli": "^5.2.0",
+    "@angular/cli": "^6.2.9",
+    "@angular/compiler-cli": "^6.1.10",
     "@angular/language-service": "^5.2.0",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
@@ -62,7 +65,7 @@
     "typescript": "~2.7.2"
   },
   "engines": {
-    "npm": "3.10.10",
-    "node": "^6.10.3"
+    "npm": "6.6.0",
+    "node": "8.12.0"
   }
 }
